requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
amqp
asgiref
attrs
billiard
cachetools
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
discord
discord.py
Django
django-allauth
frozenlist
gevent
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
joblib
kombu
multidict
networkx
numpy
oauthlib
packaging
profanity
prompt_toolkit
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pygsheets
pyparsing
python-dateutil
python-dotenv
redis
requests
requests-oauthlib
rsa
scipy
setuptools
six
smmap
sqlparse
threadpoolctl
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
yarl